move article over to tsx
